Fix go get path case sensitivity issues
`gitlab.com/TheDahv` and `gitlab.com/thedahv` are not the same thing on
disk on case-sensitive systems like Linux/OSX etc.

All the code refers to the paths correctly and the network does the
right thing resolving the username path in the request.

The documentation is the only thing that mentioned the path as "TheDahv"
so we fix this so new users don't screw up downloads.